# api/swagger.yaml
basePath: /
definitions:
...
  controller.PlayerRequestDoc:
    properties:
      api_user:
        type: string
      name:
        type: string
      universe:
        type: string
    type: object
  controller.PlayerResponseDoc:
    properties:
      api_user:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      universe:
        type: string
    type: object
...
  controller.ToolkitErrorDoc:
    properties:
      Cause: {}
      Message:
        type: string
      Value:
        type: integer
    type: object
...
info:
  contact: {}
  description: REST API for the Galactic Sovereign backend service.
  title: Galactic Sovereign API
  version: "1.0"
paths:
...
  /players:
    get:
      description: Returns players, optionally filtered by api_user.
      parameters:
      - description: API user id (UUID)
        in: query
        name: api_user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.PlayerResponseDoc'
            type: array
        "400":
          description: Invalid id syntax
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ToolkitErrorDoc'
      summary: List players
      tags:
      - players
    post:
      consumes:
      - application/json
      description: Creates a player and its homeworld.
      parameters:
      - description: Player payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.PlayerRequestDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.PlayerResponseDoc'
        "400":
          description: Invalid player syntax
          schema:
            type: string
        "409":
          description: Name already used
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ToolkitErrorDoc'
      summary: Create player
      tags:
      - players
  /players/{id}:
    delete:
      description: Deletes a player by id.
      parameters:
      - description: Player id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid id syntax
          schema:
            type: string
        "404":
          description: No such player
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ToolkitErrorDoc'
      summary: Delete player
      tags:
      - players
    get:
      description: Returns a player by id.
      parameters:
      - description: Player id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.PlayerResponseDoc'
        "400":
          description: Invalid id syntax
          schema:
            type: string
        "404":
          description: No such player
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ToolkitErrorDoc'
      summary: Get player
      tags:
      - players
...
swagger: "2.0"